Release v1.0.4: large-prime-magnitude matrix down-filter (--matrix_lp1_bound) + mid-sieve checkpoint/resume + sieve bandwidth & sqrt-seed fixes
- Add --matrix_lp1_bound: a large-prime-magnitude down-filter for
  --matrix_only replay runs that drops every large-prime-combined relation
  whose large prime exceeds the bound (and the matching partials) right after
  the relation file is loaded, without re-sieving. This lets the effective
  large-prime bound be lowered at the matrix stage to drop below the
  high-large-prime square-root cliff. Pure smooth relations are never dropped;
  the large-prime fraction is reported before and after. Default 0 (disabled).
- Add periodic, atomic, crash-safe mid-sieve checkpointing and resume
  (--checkpoint_interval, --checkpoint_batches, --checkpoint_dir, --resume).
  A killed or wall-clock-expired sieve resumes from the last checkpoint instead
  of restarting from zero. Default-off; solo and cluster-coordinator paths
  supported (workers are stateless).
- Harden sieve and autotune memory accounting: VRAM-fraction global-bucket
  sizing with headroom, 64-bit bucket arithmetic, and an autotune
  total-footprint OOM guard that rejects memory-infeasible kernel
  configurations before launch instead of failing at allocation time.
- Fix a sieve memory-bandwidth regression by moving the per-prime B_values out
  of the streamed per-prime struct into a separate device array, restoring and
  improving steady-state sieve throughput.
- Fix uint512::sqrt() returning the un-refined Newton seed, which made the SIQS
  a_target too small (up to ~2x) and reduced smooth-relation yield; it now
  seeds from above the root and returns an exact floor(sqrt(n)). Correctness
  was never affected.
- Fix hypercube dimensions above 16 silently dropping factor-base a-primes,
  which previously produced zero relations at larger sieve intervals.
- Fix an out-of-bounds access in advanceRoots for the autotune polyBlockSize=1
  probe (compute-sanitizer clean).

+++ src/postprocessing/postprocessing.cu
@@ -361,6 +361,7 @@ __device__ __forceinline__ void processCandidate(
     unsigned __int128 remainder = 0;
     bool is_valid_partial = false;
 
+
     if (!is_one) {
         bool fits_128 = true;
         #pragma unroll
@@ -641,13 +642,14 @@ __global__ void yield_prediction_kernel(
     mpqs::postprocessing::PredictionResult* __restrict__ result,
     const uint64_t* __restrict__ global_count,
     uint32_t target,
-    uint32_t total_steps,
+    uint64_t total_steps,
     const mpqs::lp::SLPPinnedStats* __restrict__ lp_stats
 ) {
     // Current relation count (lower 32 bits — count never exceeds 2^32)
     uint32_t R = static_cast<uint32_t>(*global_count);
 
-    // Yield rate: relations per sieve step
+    // Yield rate: relations per sieve step (total_steps is u64 — RSA-140 a-value counts
+    // exceed 2^32; the float division is value-preserving for the smaller-N regimes).
     float lambda = (total_steps > 0) ? static_cast<float>(R) / static_cast<float>(total_steps) : 0.0f;
 
     // LP match rate and predicted LP yield
